Create CppDriving
Project of driving game in Cpp using ASCII as GUI/Game Display.
The gist of this project is a "sliding" down towards the player game-field. Where the player is stationary at the bottom of it.
Obstacles are randomly generated and the player has to avoid them by driving left or right.